+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+2.0.2 
+
+Bugfixes:
+
+Corrected Admin User Edit
+Corrected Admin Groups Edit
+Corrected Admin New User Creation is not activated on creation
+Corrected Ldap Import Users activate login 
+
+Features:
+
+Added Reportula Rss Reader News to Administration Dashboard
+Added Reportula Logo	
+	
 2.0.1
 
 Bugfixes: